AEA-262: Translate send message payload (#62)
* AEA-262: Started writing translation for Send Message Payload.

* Moved model classes to a more appropriate module. Tidied up translation. Fixed failing tests.

* AEA-262: Refactored translation service into smaller modules.

* AEA-262: Added constants for prescription treatment type codes.

* AEA-262: Created constants for patient care provision type codes. Renamed some methods for clarity. Changed translation service to write all HL7 V3 messages in canonical form.

* AEA-262: Populate id in send message payload from the bundle identifier.

* AEA-262: Create coordinator/dist/resources as part of the build-coordinator makefile step. This is no longer created by tsc as there are no typescript files in coordinator/src/resources.

* AEA-262: Replaced an unchecked cast with a (possibly slightly nicer) typeguard.

* AEA-262: Split translation tests into separate files to reflect the code being tested.

* AEA-262: Updating failing Postman test. Correcting other errors spotted in our examples.

* AEA-262: Fixed Jest tests broken in the process of fixing Postman tests.
