--- coordinator/src/resources/ConvertWrapper.mustache
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<PORX_IN020101UK31 xmlns="urn:hl7-org:v3">
-    <id root="{{send_message_payload_id}}"/>
-    <creationTime value="20200610102631"/>
-    <versionCode code="V3NPfIT3.0"/>
-    <interactionId extension="PORX_IN020101UK31" root="2.16.840.1.113883.2.1.3.2.4.12"/>
-    <processingCode code="P"/>
-    <processingModeCode code="T"/>
-    <acceptAckCode code="NE"/>
-    <communicationFunctionRcv>
-        <device classCode="DEV" determinerCode="INSTANCE">
-            <id extension="{{to_asid}}" root="1.2.826.0.1285.0.2.0.107"/>
-        </device>
-    </communicationFunctionRcv>
-    <communicationFunctionSnd>
-        <device classCode="DEV" determinerCode="INSTANCE">
-            <id extension="{{from_asid}}" root="1.2.826.0.1285.0.2.0.107"/>
-        </device>
-    </communicationFunctionSnd>
-    <ControlActEvent classCode="CACT" moodCode="EVN">
-        <author typeCode="AUT">
-            <AgentPersonSDS classCode="AGNT">
-                <id extension="B0090,B0100,B1510" root="1.2.826.0.1285.0.2.0.67"/>
-                <agentPersonSDS classCode="PSN" determinerCode="INSTANCE">
-                    <id extension="687227875014" root="1.2.826.0.1285.0.2.0.65"/>
-                </agentPersonSDS>
-                <part typeCode="PART">
-                    <partSDSRole classCode="ROL">
-                        <id extension="S0080:G0450:R5080" root="1.2.826.0.1285.0.2.1.104"/>
-                    </partSDSRole>
-                </part>
-            </AgentPersonSDS>
-        </author>
-        <author1 typeCode="AUT">
-            <AgentSystemSDS classCode="AGNT">
-                <agentSystemSDS classCode="DEV" determinerCode="INSTANCE">
-                    <id extension="{{from_asid}}" root="1.2.826.0.1285.0.2.0.107"/>
-                </agentSystemSDS>
-            </AgentSystemSDS>
-        </author1>
-        <subject>
-            <ParentPrescription>
-            </ParentPrescription>
-        </subject>
-    </ControlActEvent>
-</PORX_IN020101UK31>
