+++ coordinator/tests/resources/test-resources.ts
@@ -1,13 +1,9 @@
 import * as XmlJs from 'xml-js'
 import * as fs from 'fs'
 import * as path from "path"
-import {ParentPrescription} from "../../src/services/hl7-v3-prescriptions";
-import {Bundle, Parameters} from "../../src/services/fhir-resources";
+import {ParentPrescription} from "../../src/model/hl7-v3-prescriptions";
+import {Bundle, Parameters} from "../../src/model/fhir-resources";
 import {ElementCompact} from "xml-js";
-
-export function clone<T>(input: T): T {
-    return JSON.parse(JSON.stringify(input))
-}
 
 const fhirMessageUnsigned1Str = fs.readFileSync(path.join(__dirname, "./parent-prescription-1/PrepareRequest-FhirMessageUnsigned.json"), "utf8")
 const fhirMessageSigned1Str = fs.readFileSync(path.join(__dirname, "./parent-prescription-1/SendRequest-FhirMessageSigned.json"), "utf8")
