+++ models/examples/example-1-repeat-dispensing/SendRequest-FhirMessageSigned.json
@@ -1,6 +1,10 @@
 {
   "resourceType": "Bundle",
   "id": "F68CEA42-E8A0-D799-C874-163048F0D390",
+  "identifier": {
+    "system": "https://tools.ietf.org/html/rfc4122",
+    "value": "A7B86F8D-1DBD-FC28-E050-D20AE3A215F0"
+  },
   "type": "collection",
   "entry": [
     {
