+++ models/examples/example-2-acute-nominated-pharmacy/PrepareRequest-FhirMessageUnsigned.json
@@ -1,6 +1,10 @@
 {
   "resourceType": "Bundle",
   "id": "9C7635DA-DE6A-38CE-B790-163048F0D390",
+  "identifier": {
+    "system": "https://tools.ietf.org/html/rfc4122",
+    "value": "A7B86F8D-1DAE-FC28-E050-D20AE3A215F0"
+  },
   "type": "collection",
   "entry": [
     {
@@ -115,14 +119,14 @@
           {
             "url": "https://fhir.nhs.uk/R4/StructureDefinition/Extension-DM-ResponsiblePractitioner",
             "valueReference": {
-              "reference": "urn:uuid:A7B86F8D-1D4A-FC28-E050-D20AE3A215F0"
+              "reference": "urn:uuid:9B4C1A77-CCCC-418A-B835-4112FBCB14F3"
             }
           }
         ],
         "identifier": [
           {
             "system": "https://fhir.nhs.uk/Id/prescription-order-item-number",
-            "value": "A7B86F8D-1D81-FC28-E050-D20AE3A215F0"
+            "value": "A7B86F8D-1D4A-FC28-E050-D20AE3A215F0"
           }
         ],
         "groupIdentifier": {
@@ -212,7 +216,7 @@
           {
             "url": "https://fhir.nhs.uk/R4/StructureDefinition/Extension-DM-ResponsiblePractitioner",
             "valueReference": {
-              "reference": "urn:uuid:DBAA616B-E23C-4D53-0F82-B2D178DB79CA"
+              "reference": "urn:uuid:9B4C1A77-CCCC-418A-B835-4112FBCB14F3"
             }
           }
         ],
@@ -309,7 +313,7 @@
           {
             "url": "https://fhir.nhs.uk/R4/StructureDefinition/Extension-DM-ResponsiblePractitioner",
             "valueReference": {
-              "reference": "urn:uuid:DBAA616B-E23C-4D53-0F82-B2D178DB79CA"
+              "reference": "urn:uuid:9B4C1A77-CCCC-418A-B835-4112FBCB14F3"
             }
           }
         ],
