+++ models/examples/example-2-acute-nominated-pharmacy/SendRequest-FhirMessageSigned.json
@@ -1,6 +1,10 @@
 {
   "resourceType": "Bundle",
   "id": "9C7635DA-DE6A-38CE-B790-163048F0D390",
+  "identifier": {
+    "system": "https://tools.ietf.org/html/rfc4122",
+    "value": "A7B86F8D-1DAE-FC28-E050-D20AE3A215F0"
+  },
   "type": "collection",
   "entry": [
     {
@@ -183,7 +187,7 @@
           {
             "url": "https://fhir.nhs.uk/R4/StructureDefinition/Extension-DM-ResponsiblePractitioner",
             "valueReference": {
-              "reference": "urn:uuid:DBAA616B-E23C-4D53-0F82-B2D178DB79CA"
+              "reference": "urn:uuid:9B4C1A77-CCCC-418A-B835-4112FBCB14F3"
             }
           }
         ],
@@ -280,7 +284,7 @@
           {
             "url": "https://fhir.nhs.uk/R4/StructureDefinition/Extension-DM-ResponsiblePractitioner",
             "valueReference": {
-              "reference": "urn:uuid:DBAA616B-E23C-4D53-0F82-B2D178DB79CA"
+              "reference": "urn:uuid:9B4C1A77-CCCC-418A-B835-4112FBCB14F3"
             }
           }
         ],
@@ -377,7 +381,7 @@
           {
             "url": "https://fhir.nhs.uk/R4/StructureDefinition/Extension-DM-ResponsiblePractitioner",
             "valueReference": {
-              "reference": "urn:uuid:DBAA616B-E23C-4D53-0F82-B2D178DB79CA"
+              "reference": "urn:uuid:9B4C1A77-CCCC-418A-B835-4112FBCB14F3"
             }
           }
         ],
