+++ models/examples/example-b-spine-error-patient-deceased/SendRequest-FhirMessageSigned.json
@@ -1,6 +1,10 @@
 {
   "resourceType": "Bundle",
   "id": "E8E8D623-62C9-4642-BC51-AB50430DF8CB",
+  "identifier": {
+    "system": "https://tools.ietf.org/html/rfc4122",
+    "value": "A7B86F8D-1DAE-FC28-E050-D20AE3A215F0"
+  },
   "type": "collection",
   "entry": [
     {
